feat: use yarn berry by default, add nodeLinker option

+++ packages/pob/lib/generators/common/husky/CommonHuskyGenerator.js
@@ -40,7 +40,7 @@ export default class CommonHuskyGenerator extends Generator {
     this.fs.delete('.git-hooks/post-checkout');
     this.fs.delete('.git-hooks/post-merge');
     this.fs.delete('.git-hooks/pre-commit');
-    this.fs.delete('.git-hooks');
+    if (this.fs.exists('.git-hooks')) this.fs.delete('.git-hooks');
 
     this.fs.delete(this.destinationPath('.huskyrc.js'));
     this.fs.delete(this.destinationPath('husky.config.js'));
